requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
requests>=2.28.0

# Qlib dependencies
qlib>=0.9.0

# Alpaca SDK
alpaca-py>=0.13.0

# Machine Learning
scikit-learn>=1.1.0
joblib>=1.2.0

# Optional ML libraries (may require OpenMP runtime)
# xgboost>=1.6.0
# lightgbm>=3.3.0

# Real-time streaming
websockets>=10.0
asyncio-mqtt>=0.11.0

# Data visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0

# Performance monitoring (optional)
psutil>=5.9.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=22.0.0
flake8>=5.0.0
mypy>=0.991 